requirements.txt
pre-commit==3.3.3
wait-for-it==2.2.0
python-dotenv==1.0.0
typing-extensions==4.5.0
pydantic==1.10.7
fastapi==0.99.1
orjson==3.8.10
uvicorn==0.22.0
httptools==0.5.0
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
backoff==2.2.1
pyjwt==2.7.0
httpx==0.24.1
python-dotenv==1.0.0
aiokafka==0.8.1
kafka-python==2.0.2
